Rename the gem pull_request_summarizer -> everypolitician-pull_request
