+++ Gemfile
@@ -1,4 +1,4 @@
 source 'https://rubygems.org'
 
-# Specify your gem's dependencies in pull_request_summarizer.gemspec
+# Specify your gem's dependencies in everypolitician-pull_request.gemspec
 gemspec
