+++ lib/everypolitician/pull_request.rb
@@ -1,0 +1,7 @@
+require 'require_all'
+require_rel 'pull_request'
+
+module Everypolitician
+  module PullRequest
+  end
+end
