+++ lib/everypolitician/pull_request/compare_popolo.rb
@@ -1,0 +1,69 @@
+require 'everypolitician/popolo'
+
+require_rel 'report'
+
+module Everypolitician
+  module PullRequest
+    class ComparePopolo
+      attr_reader :before
+      attr_reader :after
+      attr_reader :path
+
+      def self.parse(options)
+        before = Everypolitician::Popolo.parse(options[:before])
+        after = Everypolitician::Popolo.parse(options[:after])
+        new(before: before, after: after, path: options[:path])
+      end
+
+      def initialize(options)
+        @before = options[:before]
+        @after = options[:after]
+        @path = options[:path]
+      end
+
+      def people_name_changes
+        Report::PeopleNames.new(before, after).changed
+      end
+
+      def people_additional_name_changes
+        Report::PeopleAdditionalNames.new(before, after).changed
+      end
+
+      def people_added
+        Report::People.new(before, after).added
+      end
+
+      def people_removed
+        Report::People.new(before, after).removed
+      end
+
+      def wikidata_links_changed
+        Report::Wikidata.new(before, after).changed
+      end
+
+      def organizations_added
+        Report::Organizations.new(before, after).added
+      end
+
+      def organizations_removed
+        Report::Organizations.new(before, after).removed
+      end
+
+      def terms_added
+        Report::Terms.new(before, after).added
+      end
+
+      def terms_removed
+        Report::Terms.new(before, after).removed
+      end
+
+      def elections_added
+        Report::Elections.new(before, after).added
+      end
+
+      def elections_removed
+        Report::Elections.new(before, after).removed
+      end
+    end
+  end
+end
