+++ lib/everypolitician/pull_request/find_popolo_files.rb
@@ -1,0 +1,11 @@
+module Everypolitician
+  module PullRequest
+    class FindPopoloFiles
+      POPOLO_FILE_REGEX = /ep-popolo-v(\d+\.)?(\d+\.)?\d+\.json$/
+
+      def self.from(files)
+        files.select { |file| file[:filename].match(POPOLO_FILE_REGEX) }
+      end
+    end
+  end
+end
