+++ lib/everypolitician/pull_request/report/base.rb
@@ -1,0 +1,14 @@
+module Everypolitician
+  module PullRequest
+    class Report
+      class Base
+        attr_reader :before, :after
+
+        def initialize(popolo_before, popolo_after)
+          @before = popolo_before
+          @after  = popolo_after
+        end
+      end
+    end
+  end
+end
