+++ lib/everypolitician/pull_request/report/elections.rb
@@ -1,0 +1,21 @@
+module Everypolitician
+  module PullRequest
+    class Report
+      class Elections < Base
+        def elections(events)
+          events.select do |event|
+            event if event.document[:classification] == 'general election'
+          end
+        end
+
+        def added
+          elections(after.events - before.events)
+        end
+
+        def removed
+          elections(before.events - after.events)
+        end
+      end
+    end
+  end
+end
