+++ lib/everypolitician/pull_request/report/organizations.rb
@@ -1,0 +1,15 @@
+module Everypolitician
+  module PullRequest
+    class Report
+      class Organizations < Base
+        def added
+          after.organizations - before.organizations
+        end
+
+        def removed
+          before.organizations - after.organizations
+        end
+      end
+    end
+  end
+end
