+++ lib/everypolitician/pull_request/report/people.rb
@@ -1,0 +1,15 @@
+module Everypolitician
+  module PullRequest
+    class Report
+      class People < Base
+        def added
+          after.persons - before.persons
+        end
+
+        def removed
+          before.persons - after.persons
+        end
+      end
+    end
+  end
+end
