+++ lib/everypolitician/pull_request/version.rb
@@ -1,0 +1,5 @@
+module Everypolitician
+  module PullRequest
+    VERSION = '0.3.0'.freeze
+  end
+end
