--- lib/pull_request_summarizer.rb
@@ -1,6 +1,0 @@
-require 'require_all'
-require_rel 'pull_request_summarizer'
-
-module PullRequestSummarizer
-  # Your code goes here...
-end
