--- lib/pull_request_summarizer/find_popolo_files.rb
@@ -1,9 +1,0 @@
-module PullRequestSummarizer
-  class FindPopoloFiles
-    POPOLO_FILE_REGEX = /ep-popolo-v(\d+\.)?(\d+\.)?\d+\.json$/
-
-    def self.from(files)
-      files.select { |file| file[:filename].match(POPOLO_FILE_REGEX) }
-    end
-  end
-end
