--- lib/pull_request_summarizer/report/base.rb
@@ -1,12 +1,0 @@
-module PullRequestSummarizer
-  class Report
-    class Base
-      attr_reader :before, :after
-
-      def initialize(popolo_before, popolo_after)
-        @before = popolo_before
-        @after  = popolo_after
-      end
-    end
-  end
-end
