--- lib/pull_request_summarizer/report/elections.rb
@@ -1,19 +1,0 @@
-module PullRequestSummarizer
-  class Report
-    class Elections < Base
-      def elections(events)
-        events.select do |event|
-          event if event.document[:classification] == 'general election'
-        end
-      end
-
-      def added
-        elections(after.events - before.events)
-      end
-
-      def removed
-        elections(before.events - after.events)
-      end
-    end
-  end
-end
