--- lib/pull_request_summarizer/report/organizations.rb
@@ -1,13 +1,0 @@
-module PullRequestSummarizer
-  class Report
-    class Organizations < Base
-      def added
-        after.organizations - before.organizations
-      end
-
-      def removed
-        before.organizations - after.organizations
-      end
-    end
-  end
-end
