--- lib/pull_request_summarizer/report/people.rb
@@ -1,13 +1,0 @@
-module PullRequestSummarizer
-  class Report
-    class People < Base
-      def added
-        after.persons - before.persons
-      end
-
-      def removed
-        before.persons - after.persons
-      end
-    end
-  end
-end
