--- lib/pull_request_summarizer/report/terms.rb
@@ -1,19 +1,0 @@
-module PullRequestSummarizer
-  class Report
-    class Terms < Base
-      def terms(events)
-        events.select do |event|
-          event if event.document[:classification] == 'legislative period'
-        end
-      end
-
-      def added
-        terms(after.events - before.events)
-      end
-
-      def removed
-        terms(before.events - after.events)
-      end
-    end
-  end
-end
