--- lib/pull_request_summarizer/review_changes.rb
@@ -1,21 +1,0 @@
-require 'erb'
-
-module PullRequestSummarizer
-  class ReviewChanges
-    attr_reader :popolo_files
-
-    def initialize(popolo_before_after)
-      @popolo_files = popolo_before_after.map do |opts|
-        ComparePopolo.parse(opts)
-      end
-    end
-
-    def to_html
-      template.result(binding)
-    end
-
-    def template
-      @template ||= ERB.new(File.read(File.expand_path('../../../comment_template.md.erb', __FILE__)))
-    end
-  end
-end
