--- lib/pull_request_summarizer/version.rb
@@ -1,3 +1,0 @@
-module PullRequestSummarizer
-  VERSION = '0.3.0'.freeze
-end
