+++ test/everypolitician/pull_request/find_popolo_files_test.rb
@@ -1,12 +1,12 @@
 require 'test_helper'
 
-describe PullRequestSummarizer::FindPopoloFiles do
+describe Everypolitician::PullRequest::FindPopoloFiles do
   it 'returns files with a matching filename' do
     files = [
       { filename: 'countries.json' },
       { filename: 'data/UK/Commons/ep-popolo-v1.0.json' },
     ]
-    popolo_files = PullRequestSummarizer::FindPopoloFiles.from(files)
+    popolo_files = Everypolitician::PullRequest::FindPopoloFiles.from(files)
     assert_equal [{ filename: 'data/UK/Commons/ep-popolo-v1.0.json' }], popolo_files
   end
 
@@ -15,7 +15,7 @@ describe PullRequestSummarizer::FindPopoloFiles do
       { filename: 'countries.json' },
       { filename: 'data/UK/Commons/ep-popolo-v2.4.3.json' },
     ]
-    popolo_files = PullRequestSummarizer::FindPopoloFiles.from(files)
+    popolo_files = Everypolitician::PullRequest::FindPopoloFiles.from(files)
     assert_equal [{ filename: 'data/UK/Commons/ep-popolo-v2.4.3.json' }], popolo_files
   end
 end
