+++ test/everypolitician/pull_request/wikidata_changes_test.rb
@@ -1,6 +1,6 @@
 require 'test_helper'
 
-describe PullRequestSummarizer::ReviewChanges do
+describe Everypolitician::PullRequest::ReviewChanges do
   let(:before_after) do
     [
       {
@@ -45,7 +45,7 @@ describe PullRequestSummarizer::ReviewChanges do
     ]
   end
 
-  subject { PullRequestSummarizer::ReviewChanges.new(before_after) }
+  subject { Everypolitician::PullRequest::ReviewChanges.new(before_after) }
 
   it 'renders the comment template' do
     comment = subject.to_html
