--- test/pull_request_summarizer_test.rb
@@ -1,7 +1,0 @@
-require 'test_helper'
-
-describe PullRequestSummarizer do
-  it 'has a version number' do
-    ::PullRequestSummarizer::VERSION.wont_be_nil
-  end
-end
